+++ statscollectlibs/collector/StatsCollectBuilder.py
@@ -156,15 +156,15 @@ class StatsCollectBuilder(ClassHelpers.SimpleCloseContext):
 
         return stcoll
 
-    def build_stcoll_nores(self, pman, reportid, cpu=None, cmd=None, local_outdir=None,
+    def build_stcoll_nores(self, pman, reportid, cpus=None, cmd=None, local_outdir=None,
                            remote_outdir=None, local_path=None, remote_path=None):
         """
         Same as 'build_stcoll()', but does not require the caller creating a 'WORawResult' object.
-        The arguments are the same as in 'build_stcoll()', except for 'reportid', 'cpu', and
+        The arguments are the same as in 'build_stcoll()', except for 'reportid', 'cpus', and
         'cmd', which are the same as in 'StatsCollect.StatsCollect.__init__()'.
         """
 
-        self._res = WORawResult.WORawResult(reportid, local_outdir, cmd=cmd, cpu=cpu)
+        self._res = WORawResult.WORawResult(reportid, local_outdir, cmd=cmd, cpus=cpus)
         return self.build_stcoll(pman, self._res, local_outdir=local_outdir,
                                  remote_outdir=remote_outdir, local_path=local_path,
                                  remote_path=remote_path)