+++ statscollectlibs/dfbuilders/_TurbostatDFBuilder.py
@@ -4,12 +4,17 @@
 # Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: BSD-3-Clause
 #
-# Author: Adam Hawley <adam.james.hawley@intel.com>
+# Authors: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>
+#          Adam Hawley <adam.james.hawley@intel.com>
 
 """
 Provide the capability of building 'pandas.DataFrames' out of raw turbostat statistics files.
 """
 
+from __future__ import annotations # Remove when switching to Python 3.10+.
+
+from typing import Any
+from pathlib import Path
 import pandas
 from pepclibs.helperlibs.Exceptions import Error, ErrorBadFormat
 from statscollectlibs.dfbuilders import _DFBuilderBase
@@ -42,28 +47,54 @@ class TurbostatDFBuilder(_DFBuilderBase.DFBuilderBase):
     Provide the capability of building a 'pandas.DataFrames' out of raw turbostat statistics files.
     """
 
-    def _build_cols_dict(self, tstat, scope):
+    def __init__(self, cpus: list[int] | None = None):
         """
-        Build a "columns dictionary" suitable to be used with 'pandas.DataFrame.from_dict()', which
-        requires the following format: '{column_name: [value]}'. The 'tstat' dictionary has the
-        '{column_name: value}' format.
+        Initialize the class instance.
 
-        Prefix the turbostat metrics from 'tstat' with the a prefix suitable for the scope.
+        Args:
+            cpus: The measured CPU numbers.
+        """
+
+        self._cpus = cpus
+
+        # A dictionary mapping dataframe column names to the corresponding turbostat metric name.
+        # E.g., column "Totals-CPU%c1" will be mapped to 'CPU%c1'.
+        self.col2metric: dict[str, str] = {}
+
+        super().__init__("Time_Of_Day_Seconds", "TimeElapsed")
+
+    def _build_cols_dict(self,
+                         tstat: dict[str, Any],
+                         scope: str, cpu: int | None = None) -> dict[str, list[Any]]:
+        """
+        Build a columns dictionary for 'pandas.DataFrame.from_dict()'. Prefix the turbostat metrics
+        with the scope name.
+
+        Args:
+            tstat: Dictionary containing turbostat metrics in the format '{metric: value}'.
+            scope: The scope of the metrics, either "System" or "CPU".
+            cpu: The CPU number if the scope is "CPU".
+
+        Returns:
+            dict: A dictionary in the format '{column_name: [value]}' suitable for use with
+                  'pandas.DataFrame.from_dict()'.
         """
 
         cols_dict = {}
         dont_prefix_metrics = {"Time_Of_Day_Seconds", "TimeElapsed"}
 
+        if scope == "System":
+            prefix = "System"
+        elif scope == "CPU":
+            assert cpu is not None
+            prefix = f"CPU{cpu}"
+        else:
+            raise Error("BUG: unsupported scope '{scope}'")
+
         for metric, value in tstat.items():
             if metric in dont_prefix_metrics:
                 colname = metric
             else:
-                if scope == "System":
-                    prefix = "System"
-                elif scope == "CPU":
-                    prefix = f"CPU{self._cpu}"
-                else:
-                    raise Error("BUG: unsupported scope '{scope}'")
                 colname = f"{prefix}-{metric}"
 
             self.col2metric[colname] = metric
@@ -71,47 +102,62 @@ class TurbostatDFBuilder(_DFBuilderBase.DFBuilderBase):
 
         return cols_dict
 
-    def _extract_cpu_data(self, parsed_dp):
-        """Extract turbostat data for the measured CPU."""
+    def _extract_cpu_data(self, cpu: int, parsed_dp):
+        """
+        Extract turbostat data for a specific CPU number.
+
+        Args:
+            cpu: The CPU number.
+            parsed_dp: The parsed turbostat datapoint dictionary from 'TurbostatParser'.
+
+        Returns:
+            dict: A dictionary containing the turbostat data for the specified CPU.
+        """
 
         try:
-            cpu_tstat = parsed_dp["cpus"][self._cpu]
+            cpu_tstat = parsed_dp["cpus"][cpu]
         except KeyError:
-            raise Error(f"no data for CPU '{self._cpu}' found in turbostat statistics file "
-                        f"'{self._path}") from None
-
-        # Add core level values for the metrics absent at the CPU level, but present at the core
-        # level.
-        try:
-            for metric, value in parsed_dp["cpu2coreinfo"][self._cpu]["totals"].items():
-                if metric not in cpu_tstat:
-                    cpu_tstat[metric] = value
-        except KeyError:
-            raise Error(f"no data for CPU '{self._cpu}' found in turbostat statistics file "
+            raise Error(f"No data for CPU '{cpu}' found in turbostat statistics file "
                         f"'{self._path}") from None
 
         return cpu_tstat
 
-    def _turbostat_to_df(self, parsed_dp):
+    # TODO: implement proper type hint for parsed_dp and add it.
+    def _turbostat_to_df(self, parsed_dp: dict[Any, Any]) -> pandas.DataFrame:
         """
-        Convert a parsed turbostat datapoint dictionary to a dataframe. The dataframe columns will
-        be prefixed with the scope name. The arguments are as follows.
-          * parsed_dp - the parsed turbostat datapoint dictionary (from 'TurbostatParser').
+        Convert a parsed turbostat datapoint dictionary to a pandas dataframe. The DataFrame
+        columns will be prefixed with the scope name.
+
+        Args:
+            parsed_dp: The parsed turbostat datapoint dictionary from 'TurbostatParser'.
+
+        Returns:
+            pandas.DataFrame: A dataframe containing the turbostat data with appropriate column
+                   prefixes.
         """
 
         # Prepare the "columns dictionary", which has the '{colname: [value]}' format. This format
         # is required by 'pandas.DataFrame.from_dict()'. Start with the system-wide scope metrics.
         cols_dict = self._build_cols_dict(parsed_dp["totals"], "System")
-        if self._cpu is not None:
-            cpu_tstat = self._extract_cpu_data(parsed_dp)
-            cols_dict.update(self._build_cols_dict(cpu_tstat, "CPU"))
+        if self._cpus is not None:
+            for cpu in self._cpus:
+                cpu_tstat = self._extract_cpu_data(cpu, parsed_dp)
+                cols_dict.update(self._build_cols_dict(cpu_tstat, "CPU", cpu=cpu))
 
         return pandas.DataFrame.from_dict(cols_dict)
 
-    def _read_stats_file(self, path):
+    def _read_stats_file(self, path: Path) -> pandas.DataFrame:
         """
-        Return a 'pandas.DataFrame' containing the data stored in the raw turbostat statistics file
-        at path 'path'.
+        Read a raw turbostat statistics file and return its data as a pandas dataframe.
+
+        Args:
+            path: The file path to the raw turbostat statistics file.
+
+        Returns:
+            pandas.DataFrame: A DataFrame containing the data from the turbostat statistics file.
+
+        Raises:
+            ErrorBadFormat: If the raw statistics file does not include time-stamps.
         """
 
         parser = TurbostatParser.TurbostatParser(path, derivatives=True)
@@ -121,12 +167,12 @@ class TurbostatDFBuilder(_DFBuilderBase.DFBuilderBase):
             # Try to read the first data point from raw statistics file.
             parsed_dp = next(generator)
         except StopIteration:
-            raise Error(f"empty or incorrectly formatted 'turbostat' statistics file "
+            raise Error(f"Empty or incorrectly formatted 'turbostat' statistics file "
                         f"'{path}") from None
 
         # Sanity check.
         if "Time_Of_Day_Seconds" not in parsed_dp["totals"]:
-            raise ErrorBadFormat(f"rejecting turbostat statistics file '{path}' - it does not "
+            raise ErrorBadFormat(f"Rejecting turbostat statistics file '{path}' - it does not "
                                  f"include time-stamps.\nCollect turbostat statistics with the "
                                  f"'--enable Time_Of_Day_Seconds' option to include time-stamps.")
 
@@ -139,17 +185,3 @@ class TurbostatDFBuilder(_DFBuilderBase.DFBuilderBase):
             df = pandas.concat([df, next_row_df], ignore_index=True)
 
         return df
-
-    def __init__(self, cpu=None):
-        """
-        The class constructor. The arguments are as follows.
-          * cpu - the measured CPU number.
-        """
-
-        self._cpu = cpu
-
-        # A dictionary mapping dataframe column names to the corresponding turbostat metric name.
-        # E.g., column "Totals-CPU%c1" will be mapped to 'CPU%c1'.
-        self.col2metric = {}
-
-        super().__init__("Time_Of_Day_Seconds", "TimeElapsed")